Refactor SSO, link, and verification code
