+++ src/password.js
