+++ src/signon.js
@@ -5,6 +5,8 @@ import { getQueryAttr, redirectToPath } from "./url.js";
 import { exchange } from "./refresh.js";
 import { throwFormattedError } from "./utils.js";
 import { loginWithVerificationCode } from "./mfa.js";
+import { loginWithLink, sendPasswordlessLink } from "./link.js";
+import { signonWithSso } from "./sso.js";
 
 /**
  * This file has methods for signing up and logging in
@@ -40,7 +42,7 @@ export async function signup({
     case "github":
     case "google":
     case "linkedin":
-      return signupWithSSO({ provider: method, redirect });
+      return signonWithSso({ provider: method, redirect });
     case "password":
       return signupWithPassword({
         username,
@@ -57,17 +59,6 @@ export async function signup({
         'Userfront.signup called with invalid "method" property.'
       );
   }
-}
-
-/**
- * Register a new user in via SSO provider.
- * Redirect the browser after successful authentication and 302 redirect from server.
- * @param {String} provider Name of SSO provider
- */
-function signupWithSSO({ provider, redirect }) {
-  if (!provider) throw new Error("Missing provider");
-  const url = getProviderLink({ provider, redirect });
-  window.location.assign(url);
 }
 
 /**
@@ -147,7 +138,7 @@ export async function login({
     case "github":
     case "google":
     case "linkedin":
-      return loginWithSSO({ provider: method, redirect });
+      return signonWithSso({ provider: method, redirect });
     case "password":
       return loginWithPassword({
         email,
@@ -171,35 +162,6 @@ export async function login({
     default:
       throw new Error('Userfront.login called with invalid "method" property.');
   }
-}
-
-/**
- * Log a user in via SSO provider.
- * Redirect the browser after successful authentication and 302 redirect from server.
- * @param {String} provider Name of SSO provider
- * @param {String} redirect - do not redirect if false, or redirect to given path
- */
-function loginWithSSO({ provider, redirect }) {
-  if (!provider) throw new Error("Missing provider");
-  const url = getProviderLink({ provider, redirect });
-  window.location.assign(url);
-}
-
-export function getProviderLink({ provider, redirect }) {
-  if (!provider) throw new Error("Missing provider");
-  if (!store.tenantId) throw new Error("Missing tenant ID");
-
-  let url = `${store.baseUrl}auth/${provider}/login?tenant_id=${store.tenantId}&origin=${window.location.origin}`;
-
-  let redirectTo = redirect || getQueryAttr("redirect");
-  if (redirect === false) {
-    redirectTo = typeof document === "object" && document.location.pathname;
-  }
-  if (redirectTo) {
-    url += `&redirect=${encodeURIComponent(redirectTo)}`;
-  }
-
-  return url;
 }
 
 /**
@@ -256,87 +218,6 @@ async function completeSamlLogin() {
     window.location.assign(
       `${store.baseUrl}auth/saml/idp/login?tenant_id=${store.tenantId}&token=${data.token}&uuid=${store.user.userUuid}`
     );
-  } catch (error) {
-    throwFormattedError(error);
-  }
-}
-
-/**
- * Log a user in with a token/uuid combo passed into the function or
- * in the URL querystring. ?token=...&uuid=...
- * @param {String} token
- * @param {UUID} uuid
- * @param {String} redirect - do not redirect if false, or redirect to given path
- */
-export async function loginWithLink({ token, uuid, redirect } = {}) {
-  try {
-    token = token || getQueryAttr("token");
-    uuid = uuid || getQueryAttr("uuid");
-    if (!token || !uuid) return;
-
-    const { data } = await axios.put(`${store.baseUrl}auth/link`, {
-      token,
-      uuid,
-      tenantId: store.tenantId,
-    });
-
-    if (data.hasOwnProperty("tokens")) {
-      setCookiesAndTokens(data.tokens);
-      await exchange(data);
-      if (redirect === false) return data;
-      redirectToPath(
-        redirect || getQueryAttr("redirect") || data.redirectTo || "/"
-      );
-      return data;
-    }
-
-    if (data.hasOwnProperty("firstFactorCode")) {
-      return data;
-    }
-
-    throw new Error("Problem logging in.");
-  } catch (error) {
-    throwFormattedError(error);
-  }
-}
-
-/**
- * Send a login link to the provided email.
- * @param {String} email
- */
-export async function sendLoginLink(email) {
-  try {
-    const { data } = await axios.post(`${store.baseUrl}auth/link`, {
-      email,
-      tenantId: store.tenantId,
-    });
-    return data;
-  } catch (error) {
-    throwFormattedError(error);
-  }
-}
-
-/**
- * Create or update a user and send them a link to log in.
- * @param {Object} inputs
- */
-export async function sendPasswordlessLink({
-  email,
-  name,
-  username,
-  userData,
-  options,
-}) {
-  try {
-    const { data } = await axios.post(`${store.baseUrl}auth/link`, {
-      email,
-      name,
-      username,
-      data: userData,
-      options,
-      tenantId: store.tenantId,
-    });
-    return data;
   } catch (error) {
     throwFormattedError(error);
   }
