+++ test/verificationCode.spec.js
@@ -1,0 +1,6 @@
+describe("sendVerificationCode()", () => {
+  describe("sms", () => {
+    it("should send a verification code", async () => {});
+    it("should return an error", async () => {});
+  });
+});
